+++ clisops/core/subset.py
@@ -21,6 +21,8 @@ from shapely.geometry import LineString, MultiPolygon, Point, Polygon
 from shapely.ops import cascaded_union, split
 from xarray.core.utils import get_temp_dimname
 
+from clisops.utils.dataset_utils import adjust_date_to_calendar
+
 __all__ = [
     "create_mask",
     "distance",
@@ -86,6 +88,9 @@ def check_start_end_dates(func):
                 UserWarning,
                 stacklevel=2,
             )
+            kwargs["start_date"] = adjust_date_to_calendar(
+                da, kwargs["start_date"], "forwards"
+            )
             nudged = da.time.sel(time=slice(kwargs["start_date"], None)).values[0]
             kwargs["start_date"] = to_isoformat(nudged)
 
@@ -106,6 +111,9 @@ def check_start_end_dates(func):
                 '"end_date" has been nudged to nearest valid time step in xarray object.',
                 UserWarning,
                 stacklevel=2,
+            )
+            kwargs["end_date"] = adjust_date_to_calendar(
+                da, kwargs["end_date"], "backwards"
             )
             nudged = da.time.sel(time=slice(None, kwargs["end_date"])).values[-1]
             kwargs["end_date"] = to_isoformat(nudged)